исправления

+++ Jenkinsfile
@@ -24,12 +24,28 @@ pipeline {
         }
         stage('Clean image remove') {
             steps {
-                bat "docker rmi pet-api:latest"
+               script {
+                   def imageExists = sh(script: 'docker image inspect pet-api:latest > /dev/null 2>&1', returnStatus: true) == 0
+                   if (imageExists) {
+                       bat "docker rmi pet-api:latest"
+                   } else {
+                       echo "Image pet-api:latest does not exist. Skipping clean-up."
+                   }
+               }
             }
         }
         stage('Docker-compose start') {
             steps {
                 bat 'docker-compose up --build'
+                script {
+                    def containerStatus = sh(script: 'docker inspect -f {{.State.Status}} pet-api', returnStdout: true).trim()
+                    while (containerStatus != 'running') {
+                        echo "Waiting for the container to start..."
+                        sleep 10 // Подождать 10 секунд перед следующей попыткой
+                        containerStatus = sh(script: 'docker inspect -f {{.State.Status}} pet-api', returnStdout: true).trim()
+                    }
+                    echo "Container is running."
+                }
             }
         }
     }
